Make the aws-python template match aws-typescript/aws-javascript templates. (#33)
There's a small discrepancy between the `aws-python` template and the `aws-typescript` and `aws-javascript` examples.

In the Python example, the enforcement level is set on the policy pack, while in the TS/JS example, the enforcement level is set on the policy. These templates should produce the same example to avoid confusion.